--- ding/rl_utils/efficientzero/atari_env_wrapper.py
@@ -1,48 +1,0 @@
-"""
-The following code is adapted from https://github.com/YeWR/EfficientZero
-"""
-
-import numpy as np
-from ding.rl_utils.efficientzero.game import Game
-from ding.rl_utils.efficientzero.utils import arr_to_str
-
-
-class AtariWrapper(Game):
-
-    def __init__(self, env, discount: float, cvt_string=True):
-        """Atari Wrapper
-        Parameters
-        ----------
-        env: Any
-            another env wrapper
-        discount: float
-            discount of env
-        cvt_string: bool
-            True -> convert the observation into string in the replay buffer
-        """
-        super().__init__(env, env.action_space.n, discount)
-        self.cvt_string = cvt_string
-
-    def legal_actions(self):
-        return [_ for _ in range(self.env.action_space.n)]
-
-    def step(self, action):
-        observation, reward, done, info = self.env.step(action)
-        observation = observation.astype(np.uint8)
-
-        if self.cvt_string:
-            observation = arr_to_str(observation)
-
-        return observation, reward, done, info
-
-    def reset(self, **kwargs):
-        observation = self.env.reset(**kwargs)
-        observation = observation.astype(np.uint8)
-
-        if self.cvt_string:
-            observation = arr_to_str(observation)
-
-        return observation
-
-    def close(self):
-        self.env.close()
