+++ dizoo/atari/envs/atari_wrappers.py
@@ -5,6 +5,9 @@ import gym
 from collections import deque
 from ding.envs import NoopResetWrapper, MaxAndSkipWrapper, EpisodicLifeWrapper, FireResetWrapper, WarpFrameWrapper, ScaledFloatFrameWrapper, \
                         ClipRewardWrapper, FrameStackWrapper
+import numpy as np
+from ding.rl_utils.efficientzero.game import Game
+from ding.utils.compression_helper import arr_to_str
 
 
 def wrap_deepmind(env_id, episode_life=True, clip_rewards=True, frame_stack=4, scale=True, warp_frame=True):
@@ -50,7 +53,7 @@ def wrap_deepmind_mr(env_id, episode_life=True, clip_rewards=True, frame_stack=4
     :param bool warp_frame: wrap the grayscale + resize observation wrapper.
     :return: the wrapped atari environment.
     """
-    assert 'MontezumaReveng' in env_id
+    assert 'MontezumaRevenge' in env_id
     env = gym.make(env_id)
     env = NoopResetWrapper(env, noop_max=30)
     env = MaxAndSkipWrapper(env, skip=4)
@@ -67,3 +70,47 @@ def wrap_deepmind_mr(env_id, episode_life=True, clip_rewards=True, frame_stack=4
     if frame_stack:
         env = FrameStackWrapper(env, frame_stack)
     return env
+
+
+"""
+The following code is adapted from https://github.com/YeWR/EfficientZero
+"""
+class AtariWrapper(Game):
+
+    def __init__(self, env, discount: float, cvt_string=True):
+        """Atari Wrapper
+        Parameters
+        ----------
+        env: Any
+            another env wrapper
+        discount: float
+            discount of env
+        cvt_string: bool
+            True -> convert the observation into string in the replay buffer
+        """
+        super().__init__(env, env.action_space.n, discount)
+        self.cvt_string = cvt_string
+
+    def legal_actions(self):
+        return [_ for _ in range(self.env.action_space.n)]
+
+    def step(self, action):
+        observation, reward, done, info = self.env.step(action)
+        observation = observation.astype(np.uint8)
+
+        if self.cvt_string:
+            observation = arr_to_str(observation)
+
+        return observation, reward, done, info
+
+    def reset(self, **kwargs):
+        observation = self.env.reset(**kwargs)
+        observation = observation.astype(np.uint8)
+
+        if self.cvt_string:
+            observation = arr_to_str(observation)
+
+        return observation
+
+    def close(self):
+        self.env.close()
